+++ packages/core/src/index.ts
@@ -14,4 +14,5 @@ export * from "./stats.js";
 export * from "./report.js";
 export * from "./rules.js";
 export * from "./coverage.js";
+export * from "./structure/index.js";
 export * from "./classify/index.js";